upload save-log

+++ src/main/java/com/example/testSpringBoot/repository/MessageRepository.java
@@ -3,5 +3,6 @@ package com.example.testSpringBoot.repository;
 import com.example.testSpringBoot.model.Message;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MessageRepository extends JpaRepository<Message, Long> {
+// Sửa Long thành Integer để khớp với kiểu dữ liệu của messageId trong Entity
+public interface MessageRepository extends JpaRepository<Message, Integer> {
 }
